better module structure for tests
