+++ src/testframework/tests/__init__.py
@@ -1,7 +1,8 @@
-from .category import ValidCategory
-from .range import ValidNumericRange
-from .regex import IntegerString, RegexTest
-from .value import CorrectValue
+from .correct_value import CorrectValue
+from .integer_string import IntegerString
+from .regex_tst import RegexTest
+from .valid_category import ValidCategory
+from .valid_numeric_range import ValidNumericRange
 
 __all__ = [
     "RegexTest",
